+++ rancraftPenguins/EntityPenguinEmp.java
@@ -104,10 +104,12 @@ public class EntityPenguinEmp extends EntityPenguin
         return 0.4F;
     }
 
+/*  Not needed here. This function in EntityPenguin.java will set moveSpeed from this file correctly
     public void applyEntityAttributes()
     {
-        super.applyEntityAttributes(moveSpeed, MAXHEALTHTAME, MAXHEALTHWILD);
+            super.applyEntityAttributes((float) moveSpeed, MAXHEALTHTAME, MAXHEALTHWILD);
     }
+*/
 
     /**
      * (abstract) Protected helper method to write subclass entity data to NBT.
